+++ Book-Shop/src/app/components/inline-styles/inline-styles.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-inline-styles',
+  templateUrl: './inline-styles.component.html',
+  styles: [`.col-md-4{
+    background: wheat;
+  }`]
+})
+export class InlineStylesComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}